requirements.txt
google-api-python-client==1.5.1
oauth2client==3.0.0
Pillow==3.3.0
requests==2.11.0
